big integer field for user_tip

+++ maakay/models/users.py
@@ -59,7 +59,7 @@ class UserTip(models.Model):
     sender = models.ForeignKey(User, on_delete=models.CASCADE, related_name='tip_sender')
     recepient = models.ForeignKey(User, on_delete=models.CASCADE, related_name='tip_recepient')
 
-    amount = models.IntegerField()
+    amount = models.BigIntegerField()
 
     created_at = models.DateTimeField(auto_now_add=True)
     updated_at = models.DateTimeField(auto_now=True)
